GRC: move get_color to Theme.py
