+++ grc/gui/Constants.py
@@ -6,7 +6,9 @@ SPDX-License-Identifier: GPL-2.0-or-later
 
 """
 
-
+import gi
+gi.require_version('Gdk', '3.0')
+gi.require_version('Gtk', '3.0')
 from gi.repository import Gtk, Gdk
 
 from ..core.Constants import *
